# backend/ReelJett/ReelJett.WebAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ReelJett.WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Auth/ConfirmEmail": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/ForgotPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ForgotPasswordDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/ResetPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ResetPasswordDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BaseMovie/SetViewCount": {
      "post": {
        "tags": [
          "BaseMovie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BaseMovie/SetLikeCount": {
      "post": {
        "tags": [
          "BaseMovie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isLikeButton",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/AddComment": {
      "post": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieid",
            "type": "string"
          },
          {
            "in": "query",
            "name": "content",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/GetComments": {
      "get": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/LikeComment": {
      "post": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "commentid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Favourite/AddToFavourites": {
      "post": {
        "tags": [
          "Favourite"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Favourite/DeleteFromFavourites": {
      "delete": {
        "tags": [
          "Favourite"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ForgotPasswordDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "LoginDTO": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "format": "password",
          "type": "string"
        }
      }
    },
    "RegisterDTO": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "profilePhoto": {
          "type": "string"
        }
      }
    },
    "ResetPasswordDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      }
    },
...
  }
}